feat(#3742): remove redundant package
