+++ eo-parser/src/main/java/org/eolang/parser/DrErrors.java
@@ -21,8 +21,49 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
+package org.eolang.parser;
+
+import java.util.Iterator;
+import java.util.List;
+import org.cactoos.iterable.Mapped;
+import org.xembly.Directive;
+import org.xembly.Directives;
+
 /**
- * This package contains all the classes related to errors that can be found in the parsing process.
+ * Error directives.
  * @since 0.50
  */
-package org.eolang.parser.errors;
+final class DrErrors implements Iterable<Directive> {
+
+    /**
+     * Errors accumulated.
+     */
+    private final List<ParsingException> errors;
+
+    /**
+     * Ctor.
+     * @param errors The errors.
+     */
+    DrErrors(final List<ParsingException> errors) {
+        this.errors = errors;
+    }
+
+    @Override
+    public Iterator<Directive> iterator() {
+        return new org.cactoos.iterable.Joined<>(
+            new Mapped<Iterable<Directive>>(
+                error -> new Directives()
+                    .xpath("/program")
+                    .strict(1)
+                    .addIf("errors")
+                    .strict(1)
+                    .add("error")
+                    .attr("check", "eo-parser")
+                    .attr("line", error.line())
+                    .attr("severity", "critical")
+                    .set(error.getMessage()),
+                this.errors
+            )
+        ).iterator();
+    }
+}
